Implement deep-linking to worlds on the web
Add a global SceneSwitcher node, which replaces direct calls to
SceneTree.change_scene_to_file() in the menu and world code, so that
there is one central location to react to switching between worlds.

Then, when switching to a world, place the path to its directory into
the URL hash. When switching to the main menu, clear the hash. When the
game loads, check if the URL hash contains a valid world path and switch
to it if so.

This allows deep-linking to specific worlds of the game, which learners
may well want to do to share their world with their friends.

One caveat is that in some browsers sound does not work until you first
click on the game or press a key, due to security restrictions. This is
not noticeable in the normal case where the player starts on the main
menu, since the main menu has no sound, and getting to a world requires
clicking or pressing a key. But when launching straight into a world,
the music will not play immediately. The problem will resolve itself as
soon as the user presses Jump to actually start playing.

Fixes https://github.com/endlessm/candy-collective/issues/138