Define HeaderMedium style using Outward Bound font
Previously we had two label styles: the default style, using the m3x6
font (6px high); and the HeaderLarge style, with the 16px Border Basic
font.

On the main menu, the current HeaderLarge style works well: the name of
the game is quite important, and there is enough room for it. But on the
Community Worlds page, having the header take up roughly a third of the
height does not leave much room for the list of worlds.

(The Settings page currently uses the HeaderLarge style, then overriding
it to 8px. The Border Basic font used by that style is 16px so it scales
badly in the editor. At runtime it looks fine because we set the project
to use exact scaling, not anti-aliased, but it's cheating a bit.)

Add Outward Bound, a font by the same designer as Border Basic and a
natural size (7x9px) in between those two fonts, to the project. Define
a HeaderMedium label style in the theme based on it.