ExtraWorldSelector: Move header & back button out of scrolled list
Now that the “Community Worlds” heading is smaller, there is enough
space in the page for only the list of worlds to scroll.

Move the heading out of the scrollable area, leaving it at the top of
the page. Move the back button out of the scrollable area, moving it to
the bottom of the page.

Explicitly set the focus neighbour properties so that pressing "Up" from
the Back button focuses the last world (rather than the last visible
world in the list); and that pressing "Down" from the Back button
focuses the first world (rather than doing nothing). Similarly, make
pressing "Up" from the first world wrap around to the Back button at the
bottom.

Set the focus_next/_previous properties such that pressing (Ctrl+)Tab
jumps between the list of worlds and the Back button, rather than
between the buttons in the list (which can be focused with the arrow
keys or D-pad).

Fixes https://github.com/endlessm/candy-collective/issues/127