ci: Rebuild GitHub Pages when a branch is deleted
In https://github.com/endlessm/amalgamate-pages/releases/tag/v1.0.3 the
amalgamate-pages action was updated to not show builds for branches
which have been deleted. In the case where the branch was deleted due to
a PR being merged, the workflow will end up being run anyway due to the
new build on `main`, but this might happen before the branch is deleted;
and a branch may also be deleted without having been merged.

Run the workflow when a branch is deleted so that any corresponding web
build is removed.