Add background music for main menu
This is a little piece I composed/performed using strudel.cc. It roughly
follows the chord progression of the OST, but in a more relaxed mood.
It's 3 repetitions of an 8-bar phrase:

1. Harp chords (with delay); drums enter on the upbeat into bar 9
2. Drums + bass synth
3. Another repetition

The third repetition loops cleanly. The tempo is 120bpm, which means
that each repetition is 16 seconds long; thus the loop offset is set to
32 seconds from the start of the piece.

https://github.com/endlessm/candy-collective/issues/145