Add falling candy to the main menu
The progress on the candy spawner is set so that the candy falls roughly
every 2 seconds; since the background music is 120bpm, this means the
candy falls on the first beat of every bar, for quite a pleasing effect,
though it drifts out of synch after a while.

Adjust the text theme so that there is a border matching the background
colour around text, which looks better when the falling candy is behind
a word than having it be visible through the loop of an O (for example).

https://github.com/endlessm/candy-collective/issues/145